+++ crates/miden-objects/src/account/mod.rs
@@ -66,6 +66,7 @@ pub use storage::{
     PartialStorage,
     PartialStorageMap,
     StorageMap,
+    StorageMapWitness,
     StorageSlot,
     StorageSlotType,
 };