+++ crates/miden-objects/src/account/storage/mod.rs
@@ -19,7 +19,7 @@ mod slot;
 pub use slot::{StorageSlot, StorageSlotType};
 
 mod map;
-pub use map::{PartialStorageMap, StorageMap};
+pub use map::{PartialStorageMap, StorageMap, StorageMapWitness};
 
 mod header;
 pub use header::{AccountStorageHeader, StorageSlotHeader};
@@ -112,7 +112,7 @@ impl AccountStorage {
     }
 
     /// Returns a reference to the storage slots.
-    pub fn slots(&self) -> &Vec<StorageSlot> {
+    pub fn slots(&self) -> &[StorageSlot] {
         &self.slots
     }